Added support for signals
Rudamentary support, but useful for me now.  The problem is, the only safe
way to use signals is to just set a flag and check in the main line of the
code, since signals are a type of threading.  So, we have a few functions:

process.sig.catch()		- which signals to catch
process.sig.ignore()	- which signals to ignore
process.sig.default()	- which signals to set to a default action
process.sig.caught()	- did we catch a signal?

catch(), ignore() and default() can take multiple arguments, either a
sequence of signal numbers, or a table with a sequence of signal numbers:

	process.sig.catch(1,2,{3,4,{5,6},7},8)

process.caught(), if given no paramter, will return true or false if *any*
signal has been caught, or if the given signal has been caught.

	process.sig.caught()
	process.sig.caught(2)

each signal has a define as well:

	process.sig.catch(process.sig.HUP)
	process.sig.ignore(process.sig.WINCH)