Added pollsets to org.conman.net
I added the wrapper for select()/poll()/epoll() in org.conman.net because
that's the module (in my usage) that would use those functions the most.

There are three implementations, one for select(), one for poll() and one
for epoll().  All three present the same API to Lua.  It's just that some
implementations implement more options than others.  In order to determine
which one you have, you can check org.conman.net._POLL to see which
implementation is compiled in.