Bug fix---org.conman.fsys.stat() treats all-numeric file names as an integer
lua_isnumber() returns true if the item is a number, *or a string that
is a valid number*.  lua_istring() returns true if the item is a string, *or
a number*.

Darn.

I have to lose one check.  Since I rarely (if ever) pass an actual file
descriptor to org.conman.fsys.stat(), that check goes.