Bug fix---breaking out of fsys.dir() for loop would leak resources
Since I used a lightuserdata, if someone did:

	for file in fsys.dir() do
	  if file == target then
            break
          end
	end

Then the directory entry wouldn't be properly closed.  I converted the
directory related routines to a metatable, and gave it proper closure.  This
does mean that fsys.rewind(), fsys.readdir() and fsys.close() are gone, but
there are equivilents in the returned userdatatype.