Added support for Solaris for .getaffinity() and .setaffinity()
The semantics are a bit different between Linux and Solaris---Solaris only
allows binding to one CPU whereas Linux allows binding to multiple CPUs.  As
such, the Solaris version of getaffinity() returns a single number instead
of the array that the Linux version returns.  Also, the Solaris version of
setaffinity() only takes a single number for a CPUID, instead of an
open-ended number of IDs.

Remember, abstractions leak.