Make a new base64 module.  The routines in org.conman.string are now obsolete
This is because I came across some base64 encoded data that followed the
RFC-4648 standard, which is slightly different from the RFC-2045 standard
that the routines in org.conman.string.base64 use.  There are, in fact, 13
different base64 encoding standards.  Read it and weep:

http://en.wikipedia.org/wiki/Base64#Implementations_and_history

The new codebase allows you to vary the last two encoding characters, the
padding, the line length and if you should ignore characters outside the
base64 alphabet.