Cleanup---removed some references.
I realized I don't need the coroutine reference, just holding onto the
function reference is enough.  Also, I don't need to save the per-call
Lua state---just the initial one used to initialize the module is enough.

There's still some oddness with signals and coroutines (it seems the first
signal in a coroutine is lost) but then again, signals and threads don't mix
anyway, so this isn't that much of a loss, and my current usage of this
module doesn't really have to deal with signals and coroutines.

I'm still investigating this, though.