Bug fix---possible memory usage problems with fsys.gexpand()
fsys.gexpand() was written to be used in a for loop:

	for file in fsys.expand "*" do
	  print(file)
	end

but any memory consumed by the underlying call to wordexp() would not be
freed until a GC cycle, which could be awhile.  This was fixed by doing the
cleanup upon reaching the end of the loop (in the function returned by
fsys.expand() and getting rid of the whole metatable for the expansion, but
then this:

	for file in fsys.exapand "*" do
	  if file == 'foo' then break end
	end

could potentially leak actual memory.  So I wrote code to do the early
cleanup with fsys.gexpand() but also protect against the early clean up
never happening.