Refactor the process limit code.
I wanted to support Lua 5.3 and it's __pairs metamethod on the resource
limits.  In doing so, I found a much better way of dealing with the process
limits enum values.  There are some changes to return values---I no longer
return +inf for large values, but -1.  Yes, it's a breaking change.