Bug fix---dealing with leaky abstractions
There is no consistency with any documentation for select()---some say the
'exceptsfd' field is for errors, others for urgent TCP data.  Also, the with
poll(), you can't select for POLLHUP or POLLERR or any of those---they're
just set.  The same is for epoll().  So in the end, I really only have the
following three flags:

	r	- read event
	w	- write event
	p	- priority/error event

with only 'r' and 'w' being consistent.  So that's what I'm going with.