Bug fix---tls.handshake() worked differently than I expected.
It's an API change (sigh).  But I now know how the handshake works, and it
works fine in event-mode so that's good.  It works more like
tls.read()/tls.write() in that you need to call back (or yield) when you get
WANT_INPUT or WANT_OUTPUT results from tls_handshake()---it's not a boolean
result like I thought it was.