Bug fix---large data transfers don't work in TLS
Which is now fixed.  The problem was that I was feeding data into the socket
faster than it could be transmitted out.  It took awhile, but I finally got
the framework to pause when that happens.  It also happens that when TLS
detects too much data being written, it returns tls.ERROR to the upper
layers, so I need to track when I do a shortwrite at the lower layers so I
know when I need to pause and when I get an actual error.