Add traceback when the TLS context is NULL
I suspect the connection is being garbage collected before we get to writing
the data, but I don't know how that can happen.  Hopefully, by getting a
traceback in Ltls___gc() and Ltls_write() we can figure that out.  I'm
testing the hypothosis by saving the traceback in Ltls___gc(), and printing
it out (if *tls is NULL).