Bug fix---dealing with socket errors in NFL
After a few issues were the CPU goes to 100% because of an error with
reading/writing a socket, I'm chaning how I deal with it globally---close
the connection, set EOF and schedule the waiting coroutine.  The coroutine
will get the error, or at least the EOF, and things should clean up nearly
afterwards.  I hope.