Remove org.conman.net.interfaces()
I'm rethinking this function call.  It's Linux specific, and this codebase
runs generally on POSIX systems.  So I feel like removing this is in order
(especially since I don't use it, and I'm not sure who else uses it).  I'm
willing to remove this now, before I "release" this via LuaRocks.