Bug fix---running a dead coroutine
The comment (now deleted) asked "why [is this] being triggered?"---this
being a dead coroutine in the runqueue.  This can happen if one coroutine
schedules a coroutine that has since finished (or crashed) by the time
it's scheduled to run again.

A case for this is as follows:  A coroutine needs to spawn a few coroutines
to do some work, say, two concurrent requests, but we've set up a timeout
condition on the "primary" coroutine.  If either (or both) of the requests
take too long, the primary coroutine times out and runs to completion, but
the child coroutines do finish and schedule the primary to run, which has
since stopped running.

The entire code block, except for the syslog(), can go as the ref count has
already been taken care of.