Bug fix---mismatch between Lua and C.
The original Lua code had the following:

	if breakhere then ...

In Lua, 0 isn't falsy, only nil and false are, so a value of 0 for breakhere
was valid.  In the C translation, I treated a value of 0 for breakhere as
"no break at all", which in one corner case lead to an out-of-memory error.

Man, this was a tough bug, but in retrospect, a bit of fun to track down.